requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
h11==0.16.0
idna==3.11
pydantic==2.12.0
pydantic_core==2.41.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.37.0
fastapi==0.119.0
uvicorn[standard]==0.37.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.30.0
python-dotenv==1.0.0
pydantic==2.12.0
